Refactor story decorators
